release/2.0.0-dev
See release notes for mazeltov-project v2.0.0. This is the core
version meant to be compatible with this release. The big
change here is that services, controllers, and models have switched
to a system using loaders. Loaders can be configured to work in a
variety of ways, but the end goal is to allow exporting/loading
code in a way that reduces mistakes/boilerplate.

Other notable changes
- util library moved from separate repo into here
- http routers now wrappable with controllers to allow passing of
  services and models and logger automatically. These controllers
  have an almost identical API to express routers, but allow the
  use of onRedux 'xWebHttpRouterX' style hooks to modify the
  routers stack upstream.
- Many more changes to core views and layouts including
  * A dev debug dump at bottom of page
  * Menu system is now redux-able